Reorganize the existing handlers for a app scale easily
