+++ src/bot/handlers/__init__.py
@@ -1,0 +1,2 @@
+from handlers.echo import echo_handler  # noqa
+from handlers.start import start_handler  # noqa
