+++ src/bot/handlers/start.py
@@ -1,7 +1,6 @@
 from telegram import Update
-from telegram.ext import CommandHandler, ContextTypes, MessageHandler, filters
+from telegram.ext import CommandHandler, ContextTypes
 
-from core.logging import log_errors
 
 async def start(update: Update, context: ContextTypes.DEFAULT_TYPE):
     """/start command handler."""
@@ -13,15 +12,4 @@ async def start(update: Update, context: ContextTypes.DEFAULT_TYPE):
         ),
     )
 
-
-@log_errors
-async def echo(update: Update, context: ContextTypes.DEFAULT_TYPE):
-    """Echo messages handler."""
-    await context.bot.send_message(
-        chat_id=update.effective_chat.id,
-        text=update.message.text,
-    )
-
-
 start_handler = CommandHandler('start', start)
-echo_handler = MessageHandler(filters.TEXT & (~filters.COMMAND), echo)
